Interface entity
